requirements.txt
fastapi
uvicorn
pydantic
requests
playwright
google-generativeai
gunicorn